third_party/WebKit/LayoutTests/http/tests/serviceworker/install-phase-event-waituntil.html

<!DOCTYPE html>
<title>InstallPhaseEvent: waitUntil</title>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script src="resources/test-helpers.js"></script>
<script>
function runTest(test, scope, onRegister) {
  var script = 'resources/install-phase-event-waituntil.js';
  test.step(function() {
      service_worker_unregister_and_register(test, script, scope)
        .then(function(registration) {
            return wait_for_update(test, registration);
          })
        .then(test.step_func(onRegister));
    });
}

function syncWorker(test, worker, obj) {
  var channel = new MessageChannel();
  channel.port1.onmessage = test.step_func(function(e) {
      var message = e.data;
      assert_equals(message, 'SYNC',
                    'Should receive sync message from worker.');
      obj.synced = true;
      channel.port1.postMessage('ACK');
    });
  worker.postMessage({port: channel.port2}, [channel.port2]);
}

async_test(function(t) {
    // Passing scope as the test switch for worker script.
    var scope = 'install-fulfilled';
    var onRegister = function(worker) {
        var obj = {};
        worker.onstatechange = t.step_func(function() {
            if (worker.state == 'installing') {
              syncWorker(t, worker, obj);
            } else if (worker.state == 'installed') {
              assert_true(
                  obj.synced,
                  'state should be "installed" after the waitUntil promise ' +
                      'for "oninstall" is fulfilled.');
              service_worker_unregister_and_done(t, scope);
            }
          });
      };
    runTest(t, scope, onRegister);
  }, 'Test install event waitUntil fulfilled');

async_test(function(t) {
    var scope = 'activate-fulfilled';
    var onRegister = function(worker) {
        var obj = {};
        worker.onstatechange = t.step_func(function() {
            if (worker.state == 'activating') {
              syncWorker(t, worker, obj);
            } else if (worker.state == 'activated') {
              assert_true(
                  obj.synced,
                  'state should be "activated" after the waitUntil promise ' +
                      'for "onactivate" is fulfilled.');
              service_worker_unregister_and_done(t, scope);
            }
          });
      };
    runTest(t, scope, onRegister);
  }, 'Test activate event waitUntil fulfilled');

async_test(function(t) {
    var scope = 'install-rejected';
    var onRegister = function(worker) {
        worker.onstatechange = t.step_func(function() {
            if (worker.state == 'redundant')
              service_worker_unregister_and_done(t, scope);
          });
      };
    runTest(t, scope, onRegister);
  }, 'Test install event waitUntil rejected');

async_test(function(t) {
    var scope = 'activate-rejected';
    var onRegister = function(worker) {
        worker.onstatechange = t.step_func(function() {
            if (worker.state == 'redundant')
              service_worker_unregister_and_done(t, scope);
          });
      };
    runTest(t, scope, onRegister);
  }, 'Test activate event waitUntil rejected.');

async_test(function(t) {
    var scope = 'activate-multiple-fulfilled';
    var onRegister = function(worker) {
        var obj1 = {};
        var obj2 = {};
        worker.onstatechange = t.step_func(function() {
            if (worker.state == 'activating') {
              syncWorker(t, worker, obj1);
              syncWorker(t, worker, obj2);
            } else if (worker.state == 'activated') {
              assert_true(
                  obj1.synced && obj2.synced,
                  'state should be "activated" after all waitUnitl promises ' +
                      'for "onactivate" are fulfilled.');
              service_worker_unregister_and_done(t, scope);
            }
          });
      };
    runTest(t, scope, onRegister);
  }, 'Test InstallPhaseEvent multiple waitUntil fulfilled.');

async_test(function(t) {
    var scope = 'activate-reject-precedence';
    var onRegister = function(worker) {
        worker.onstatechange = t.step_func(function() {
            if (worker.state == 'redundant')
              service_worker_unregister_and_done(t, scope);
          });
      };
    runTest(t, scope, onRegister);
  }, 'Test InstallPhaseEvent waitUntil reject precedence.');
</script>
